Move ConfigurationFactory to Domain layer
